backend/requirements.txt
# GuitarLab - Backend Dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
python-dotenv==1.0.1
google-generativeai==0.8.4
chromadb==0.5.23
pydantic==2.10.4
pydantic-settings==2.7.1
Pillow==11.1.0
aiofiles==24.1.0
sqlite-utils==3.38
jinja2==3.1.5
httpx==0.28.1
